+++ include/SNUDC_LinkDef.h
@@ -1,0 +1,5 @@
+#if defined(__ROOTCLING__)
+#pragma link C++ class SNUDCTree+;
+#pragma link C++ class AnalyzerCore+;
+#pragma link C++ class SBAnalyzer+;
+#endif
