+++ src/SBAnalyzer.C
@@ -1,30 +1,16 @@
-#define SNUDC_cxx
-#include "SNUDC.h"
-#include "mylib.h"
-#include "canvas_margin.h"
-#include <TH2.h>
-#include <TStyle.h>
-#include <TCanvas.h>
+#ifndef SBAnalyzer_cxx
+#define SBAnalyzer_cxx
 
-TString WORKING_DIR = getenv("SNUDC_WORKING_DIR");
-TString root_file_path = WORKING_DIR+"/rootfiles/data/";
-TString root_filename = "runA_partial.root";
-TString filename = root_file_path + root_filename;
+#include "SBAnalyzer.h"
 
-SNUDC::SNUDC(TTree *tree) : fChain(0)
-{
-  if (tree == 0) {
-    TFile *current_file = new TFile ((filename)) ;
-    if (!current_file || !current_file->IsOpen()) {
-      current_file = new TFile(filename);
-    }
-    current_file->GetObject("SNUDC",tree);
-    
-  }
-  Init(tree);
+SBAnalyzer::SBAnalyzer(){
+  TString WORKING_DIR = getenv("SNUDC_WORKING_DIR");
+  TString filename=WORKING_DIR+"/rootfiles/data/runA0.root";
+  fChain->Add(filename);
+  Init();
 }
 
-void Draw_Hist(TString histname){
+void SBAnalyzer::Draw_Hist(TString histname){
 
   TString ENV_PLOT_PATH = getenv("SNUDC_PLOT_PATH");
   TString base_plotpath = ENV_PLOT_PATH + "/";
@@ -34,12 +20,13 @@ void Draw_Hist(TString histname){
   canvas_margin(c1);
   c1->cd();
 
-  maphist_TH1D[histname] -> Draw("hist");
+  TH1* hist=GetHist(histname);
+  if(hist) hist->Draw("hist");
 
   c1 -> SaveAs(base_plotpath + "/" + histname + ".pdf");
 }
 
-void Draw_Hist_Cumulative(TString histname){
+void SBAnalyzer::Draw_Hist_Cumulative(TString histname){
 
   TString ENV_PLOT_PATH = getenv("SNUDC_PLOT_PATH");
   TString base_plotpath = ENV_PLOT_PATH + "/";
@@ -52,14 +39,17 @@ void Draw_Hist_Cumulative(TString histname){
   gStyle->SetStatH(0.1);
   canvas_margin(c2);
   c2->cd();
-  TH1* hc =maphist_TH1D[histname] ->GetCumulative();
-  hc -> Draw("hist");
+  TH1* hist=GetHist(histname);
+  if(hist){
+    TH1* hc =hist->GetCumulative();
+    hc -> Draw("hist");
+  }
   c2 -> SaveAs(base_plotpath + "/" + histname + "_Cumulative.pdf");
 
 }
 
 
-void SNUDC::Loop()
+void SBAnalyzer::Loop()
 {
 //   In a ROOT session, you can do:
 //      root> .L SNUDC.C
@@ -86,7 +76,7 @@ void SNUDC::Loop()
 //by  b_branchname->GetEntry(ientry); //read only this branch
    if (fChain == 0) return;
 
-   Long64_t nentries = fChain->GetEntriesFast();
+   Long64_t nentries = fChain->GetEntries();
    //Long64_t nentries = 2;
    
    Long64_t nbytes = 0, nb = 0;
@@ -96,23 +86,11 @@ void SNUDC::Loop()
       nb = fChain->GetEntry(jentry);   nbytes += nb;
       // if (Cut(ientry) < 0) continue;
       
-      if(jentry%2000 == 0) cout << jentry << "/" << nentries << " done" << endl;
+      if(jentry%100 == 0) cout << jentry << "/" << nentries << " done" << endl;
 
-      for(int i = 0; i < 192; i++){
-	if(wires.at(i)->size() > 0){
-	  if(i < 16) FillHist("x1_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=16 && i < 32) FillHist("u1_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=32 && i < 48) FillHist("v1_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=48 && i < 64) FillHist("x2_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=64 && i < 80) FillHist("u2_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=80 && i < 96) FillHist("v2_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=96 && i < 112) FillHist("x3_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=112 && i < 128) FillHist("u3_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=128 && i < 144) FillHist("v3_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=144 && i < 160) FillHist("x4_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=160 && i < 176) FillHist("u4_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else if(i >=176 && i < 192) FillHist("v4_TDC", wires.at(i)->at(0), 1., 3000, 0., 3000.);
-	  else continue;
+      for(int i = 0; i < NWIRES; i++){
+	if(wires[i]->size() > 0){
+	  FillHist(GetWireName(i)(0,2)+"_TDC", wires[i]->at(0), 1., 3000, 0., 3000.);
 	}
       }
    }
@@ -126,3 +104,5 @@ void SNUDC::Loop()
    cout << "nentries : " << nentries << endl;
    cout << "nb : " << nb << endl;
 }
+
+#endif
